Аадптация разработки под 4.7 [Rx]
# Conflicts:
#	GD.MainSolution/GD.MainSolution.Shared/Sungero.Docflow/Module.mtd
#	GD.MainSolution/GD.MainSolution.Shared/Sungero.RecordManagement/Module.mtd

+++ GD.MainSolution/GD.MainSolution.Shared/Sungero.RecordManagement/ModuleSharedFunctions.cs
@@ -20,8 +20,8 @@ namespace GD.MainSolution.Module.RecordManagement.Shared
     [Public]
     public override void SynchronizeAttachmentsToActionItem(Sungero.Docflow.IOfficialDocument primaryDocument,
                                                             List<Sungero.Content.IElectronicDocument> addenda,
-                                                            List<int> addedAddendaIds,
-                                                            List<int> removedAddendaIds,
+                                                            List<long> addedAddendaIds,
+                                                            List<long> removedAddendaIds,
                                                             List<Sungero.Domain.Shared.IEntity> otherAttachments,
                                                             Sungero.RecordManagement.IActionItemExecutionTask actionItem)
     {
